Refactoring for multi-variable declaration
Chapter 39, get ready to codegen variable initialization

+++ mine/lib/mylib.c
@@ -2,7 +2,7 @@
 #include <unistd.h>
 #include <errno.h>
 void printint(long x) {
-  printf("%ld\n", x);
+  printf("%d\n", (int)x);
 }
 
 void printchar(long x) {
